pyproject.toml
[project]
name = "deploy-and-manage-adk-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage>=2.19.0",
    "locust==2.37.10",
    "networkx>=3.5",
    "pydantic>=2.11.7",
    "python-dotenv==1.1.0",
    "floggit>=0.0.19",
    "fastapi[standard]>=0.117.1",
    "google-cloud-spanner>=3.59.0",
    "google-adk>=1.18.0",
]

[dependency-groups]
dev = [
    "pytest==8.4.0",
    "ruff==0.11.13",
]
